Finished exercise 08 in JS exercises
At first I was using for loops to check for the existence of the
element in args and then splice it with indexOf to get the index.
This proved cumbersome and also failed the test that required you
to remove multiple instances of the same element. After some
research I found that the filter method worked really well for a
one line solution, where we iterate through the given arr, check
to see if the element is in the args array and if it isn't, include
it in the filtered array. Found when checking the solution that mine
was the more advanced one, which I'm proud of.