Finished exercise 09 of JS exercises
The actual loop to sum was simple. I implemented a check to switch
the start and end values with array destructuring to avoid having
to flip the loop. After seeing the tests, I implemented checks in
one if statement to check whether the values are positive integers.
The solution did it in two if statements, but I thought one with the
or operator would make more sense. Splitting the checks might help
with performance, perhaps?