Finished exercise 10 of JS exercises
Figured that checking if the year was divisible by 4 first and then
checking to see if it was a century and it wasn't divisible by 400
to return false would make sense, returning true if the if statement
was falsy. I also had a fall back that returned false if the first
condition wasn't met. Realized in the solution that you can use
parentheses for logic within the same if statement, and checking
whether it's divisible by 4 AND is not a century OR is divisible
by 400 prevented nested if statements. I didn't use variables like
the solution, but they helped understand the logic behind it.